lookup_service/pyproject.toml
[project]
name = "lookup-service"
version = "1.0.0"
description = "lookup service for quotes, contacts and products"
authors = [
  { name = "Rahul Radhakrishnan", email = "imrahul.r1795@gmail.com" }
]
dependencies = [
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.115.11",
    "opentelemetry-distro>=0.52b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.0",
    "opentelemetry-instrumentation-fastapi>=0.52b0",
    "opentelemetry-instrumentation-logging>=0.52b0",
    "structlog>=25.2.0",
    "uvicorn>=0.34.0",
    "httpx"    
]
